refactor: move DiffOption, WorktreeInfo, GitContext to @plannotator/shared
Eliminates duplicate interface definitions across server and client.
Types now live in packages/shared/types.ts and are imported by both
packages/server/git.ts and packages/review-editor/.
