+++ packages/review-editor/package.json
@@ -7,6 +7,7 @@
     "./styles": "./index.css"
   },
   "dependencies": {
+    "@plannotator/shared": "workspace:*",
     "@plannotator/ui": "workspace:*",
     "@pierre/diffs": "^1.1.0-beta.19",
     "highlight.js": "^11.11.1",
